fix:arrange the runtime exception to MessageService

+++ src/main/java/Service/MessageService.java
@@ -1,12 +1,15 @@
 package Service;
 
 import Model.Message;
+import Model.User;
 import Repository.MessageRepository;
+import Repository.UserRepository;
 import lombok.*;
 import org.springframework.stereotype.Service;
 
 import java.sql.SQLException;
 import java.util.List;
+import java.util.Optional;
 
 
 @AllArgsConstructor
@@ -14,26 +17,30 @@ import java.util.List;
 @Service
 public class MessageService {
     private MessageRepository messageRepository;
+    private UserRepository userRepository;
     public List<Message> displayMessagesBetweenTwoUser(int sender, int receiver) throws SQLException {
         if(sender == receiver){
-            System.out.println("the id cannot be the same");
+            throw new IllegalArgumentException("the id cannot be the same");
         }
         return messageRepository.AllMessageBetweenTwoUsers(sender,receiver);
     }
     public void sendMessage(Message insert) throws SQLException {
         if(insert.getAccountSender() == insert.getAccountReceiver()){
-            System.out.println("the id cannot be the same");
+            throw new IllegalArgumentException("the id cannot be the same");
         }
         messageRepository.insertNewMessage(insert);
     }
    public void UpdateSeenDatetime(int sender, int receiver) throws SQLException {
        if(sender == receiver){
-           System.out.println("the id cannot be the same");
+           throw new IllegalArgumentException("the id cannot be the same");
        }
        messageRepository.updateById(sender,receiver);
    }
    public void deleteMessage(int idUser) throws SQLException{
-
+       Optional<User> existingUserOptional = userRepository.findById(idUser);
+       if(existingUserOptional.isEmpty()){
+           throw new ResourceNotFoundException("User not found with id"+idUser);
+       }
        messageRepository.deleteMessageById(idUser);
    }
 }
